pyproject.toml
[project]
name = "budget-router"
version = "0.1.0"
description = "Incident Commander for Budgeted Tool/API Reliability — OpenEnv RL Environment"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Akshay Babbar" },
]
dependencies = [
    "pydantic>=2.0,<3.0",
    "openenv-core>=0.1.0",
    "openai>=1.0,<2.0",
    "typer>=0.12,<1.0",
    "uvicorn>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["budget_router/tests"]
pythonpath = ["."]


